+++ tadrep/characterize.py
