+++ tadrep/detect.py
